% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denumerator.R
\name{apply_to_results}
\alias{apply_to_results}
\title{Apply \code{fun} to all results out of a deseq object.}
\usage{
apply_to_results(deseq_object, results, fun, ...)
}
\arguments{
\item{deseq_object}{A Deseq Dataset with results (after the DESeq call)}

\item{results}{A list of result names (as from \code{DESeq2::listNames}), for each
result on which the function should be applied to.}

\item{fun}{A function to be applied}

\item{...}{Further arguments to be passed to \code{fun}}
}
\value{
A list with one frame per item in \code{results}, with the output of calling
\code{fun} on that result.
}
\description{
Apply \code{fun} to all results out of a deseq object.
}
