% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{count_data}
\alias{count_data}
\title{Example count data taken from a real experiment but anonimized trough random
jittering of the counts.}
\format{
An object of class \code{data.frame} with 37040 rows and 12 columns.
}
\usage{
count_data
}
\description{
The dataset has samples as columns and genes as rows.
}
\keyword{datasets}
