% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{count_metadata}
\alias{count_metadata}
\title{Example metadata of the count data from a real experiment but anonimized.
Each column is a different experimental condition and each row is a column
in the count data.}
\format{
An object of class \code{data.frame} with 12 rows and 2 columns.
}
\usage{
count_metadata
}
\description{
Example metadata of the count data from a real experiment but anonimized.
Each column is a different experimental condition and each row is a column
in the count data.
}
\keyword{datasets}
