% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeNumerator-package.R
\docType{package}
\name{DeNumerator-package}
\alias{DeNumerator}
\alias{DeNumerator-package}
\title{DeNumerator: Denumerate A DESeq2 Object To See It Better}
\description{
Inspect your DESeq2 analysis results, especially with complicated design matrices, as a nice plot.
}
\author{
\strong{Maintainer}: Luca Visentin \email{luca.visentin@unito.it} (\href{https://orcid.org/0000-0003-2568-5694}{ORCID})

}
\keyword{internal}
