% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denumerator.R
\name{plot_enumeration_frame}
\alias{plot_enumeration_frame}
\title{Plot an enumeration frame from a list of enumeration data}
\usage{
plot_enumeration_frame(
  enum_data,
  order_by = c("frequency", "label"),
  zero_lab = "zero",
  pos_lab = "positive",
  neg_lab = "negative",
  title = NULL,
  top_n = NULL,
  labels_y_nudge = 0,
  positive_symbol = "+",
  negative_symbol = "-",
  symbol_size = 10,
  positive_color = "red",
  negative_color = "blue",
  symbol_fontface = "bold",
  exclude_all_negative = FALSE,
  category_renames = NULL
)
}
\arguments{
\item{enum_data}{A list with one slot per enumeration, with items returned by
\code{to_enumeration_vector}.}

\item{order_by}{Order of the rows in the resulting plot. Either "frequency", to
sort by frequency, or "label" to sort by categorical labels (up, zero, down).}

\item{zero_lab}{Label in the enumeration for non-differentially expressed genes.}

\item{pos_lab}{Label in the enumeration for upregulated genes.}

\item{neg_lab}{Label in the enumeration for downregulated expressed genes.}

\item{title}{Title of the plot.}

\item{top_n}{If not NULL, an integer. Keep only \code{top_n} rows by frequency.}

\item{labels_y_nudge}{By default, the symbols are centered on the correct
position. However, based on the used font, the symbols might be misaligned
in the Y axis. The "nudge" to give to each symbol is plot size and
font dependent, so it's difficult to compute beforehand.
Tweak this parameter to move the symbols up (e.g. +0.02) or down (e.g. -0.02)
so that they are perfectly aligned for your specific usecase.}

\item{positive_symbol}{The symbol in the plot to give to the upregulated case.}

\item{negative_symbol}{The symbol in the plot to give to the negative case.}

\item{symbol_size}{The size of the positive and negative legend symbols.}

\item{positive_color}{The color in the plot to give to the upregulated case.}

\item{negative_color}{The color in the plot to give to the negative case.}

\item{symbol_fontface}{The font face of the positive and negative symbols.}

\item{exclude_all_negative}{If TRUE, the case where all variables are zero is
omitted from the plot.}

\item{category_renames}{A list of \code{old_label = new_label} with the new names
of the enumeration categories. The old labels are the column names of the
input enumeration data.}
}
\value{
A \code{ggproto} object with the resulting plot.
}
\description{
This function takes care of computing frequency tables from input enumerations
and plotting the resulting frequency graph.
}
\author{
Hedmad
}
