% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denumerator.R
\name{to_enumeration_vector}
\alias{to_enumeration_vector}
\title{Convert a DESeq2 "result" object to an enumeration object}
\usage{
to_enumeration_vector(
  data,
  alpha = 0.01,
  fold_change = 1,
  zero_lab = "zero",
  pos_lab = "positive",
  neg_lab = "negative",
  colname = "class"
)
}
\arguments{
\item{data}{A DESeq2 object from the 'results' function}

\item{alpha}{The alpha p-value threshold above which a gene is considered
to be not differentially expressed. E.g. 0.05 would cause any gene with a
p-value of 0.06 to be not differentially expressed.}

\item{fold_change}{The absolute fold change value a gene must have to be
considered to as differentially expressed. E.g. 2 would consider a gene
with fold change 1 to be not deferentially expressed.}

\item{zero_lab}{Label to give in the enumeration to non-differentially expressed genes.}

\item{pos_lab}{Label to give in the enumeration to upregulated genes.}

\item{neg_lab}{Label to give in the enumeration to downregulated expressed genes.}

\item{colname}{The column name to give to the variable with the labels in the resulting frame}
}
\value{
A data.frame with one row per input gene and a column with the proper
enumeration labels based on the filter in the column "colname", by default
"class".
}
\description{
Convert a DESeq2 "result" object to an enumeration object
}
